basic provider-definition and provider spec classes
tests green for load/dump operations on config.
defaults on config spec for LeapServiceProvider should
be outputting a minimal working config for bootstrapping
the connection.

we should be using this config mechanism for other config
operations; but probably we should specify a local_editable flag
for those configs that are only changed by server.