more generic error handler in EipConductorAppMixin
documentation of the Exception Hierarchy and attributes.

also a bit of general cleanup around error handling in conductor.
Hopefully to be polished an abstracted to leap.base with time.

not all errors are converted (and the old with_errors/ignoring
errors) are still there, but we should be using this style of handlers
from now on.

wrapping up with this pseudo-feature for now. as we work on individual
features we can mimick the exceptions that are working.