removed branding info from branding config file
So we officially can say this is a generic client now.
Branding config file is still at pkg/branding/config.py

Note that with this change the package is called now leap-client
so you will have to remove old installs of leap-foo-client in your
path (from previous branded builds).

Changing the package name was an AWFUL and painful idea, and we will not
do that again. (launcher is another story). Lesson learned.