fix versioneer build and sdist commands
It turns out that I was using remnants of old functions in which
I had to hook the branding mechanism, but that's not needed anymore, so
it's much simple.

With this change, we get the correct _version generated after a build or
sdist command.

For the bundles, we should use that _version. Ideally we should use the
whole build tree for bundles instead of the source tree directly.