Improve error handling during soledad boostrapping in the client.
The aim is to have better logs for debugging the different problems
behind issues like #3619 and #3867.

As I'm finding a good quantity of SSL handshake timeouts, I'm also
adding a litte retry subroutine to the load_and_sync.

Also, initialization and sync calls are separeted to be able to
correlate logs with server-side soledad.