workaround for using keyring inside a virtualenv
(securestorage backend)

This is related to research #4190 and #4083.
I didn't resolve them, but I ended understanding a bit more what kind
of issues can we be having with those.
This workaround is more than anything a cleanup on that for future work,
and is making me able to test the client much more nicely inside a
virtualenv (where the default keyring selecting was the plaintext one).

For this to work inside a virtualenv, one have to install SecureStorage,
which in turns depends on python-dbus, which is basically uninstallable
using pip inside a venv. So, symlinking to the rescue!

it needs gi/repository and pyobject to be symlinked too.
but at least you don't have to type the pass every time.

I don't know what should be do in the long term with this issue.
Some of us were not too fond on using the keyrings at all.
We don't have many options among all the keyring backends, and sadly
many of them depend on PyCrypto. So probably roll our own backend. Yay.