Grep for gnome-shell as it were a polkit agent.
The $DESKTOP_SESSION is not reliable since is set to 'default' in any
case that you choose as your default desktop environment.

Also, after some tests, I've seen that the `gnome-shell` process is not
launched in the fallback mode, so we use it as a condition for the
polkit agent present.