Disable user/pass remembering temporarily.
Set the keyring to None in order to simulate an always unavailable
keyring, that way we avoid the possibility of the user running into the
existing keyring issues. See https://leap.se/code/issues/4190

Update comparisons to do a proper comparison with `None`.
Fix login widget 'enabled' changer in order to change the 'remember'
widget *only* if we have an usable keyring.