Changed build-depends for python-requests to resolve the following lintian issue:
W: bitmask source: build-depends-on-1-revision build-depends: python-requests (>= 1.2.3-1)

N:    The package declares a build dependency on a version of a package with a
N:    -1 Debian revision such as "libfoo (>= 1.2-1)". Such a dependency will
N:    not be satisfied by a backport of libfoo 1.2-1 and therefore makes
N:    backporting unnecessarily difficult. Normally, the -1 version is
N:    unneeded and a dependency such as "libfoo (>= 1.2)" would be sufficient.
N:    If there was an earlier -0.X version of libfoo that would not satisfy
N:    the dependency, use "libfoo (>= 1.2-1~)" instead.