[bug] handle disabled registration, error 403.
If the user wants to register a new account we check whether the
provider allows registration or not right after getting the
provider.json file and show an error msg on the wizard if not allowed.

Also, there is a new signal to handle the error raised by the server if
a registration attempt is made but is rejected with error 403.

- Resolves: #6594