[bug] move events server init to backend
If the events server is initialized in a different process than the backend,
the txzmq socket raises an "zmq.error.ZMQError: Interrupted system call"
exception during the events server initialization. Despite that, communication
seems to work flawlessly after the initialization.

Moving the events server initialization to the same process as the backend
causes the exception to not be raised during events server intialization.