[bug] do not unset soledad proxy object too early
The backend was unsetting the soledad proxied object before actually calling
the soledad close() method and this was causing the application to hang when
exitting.

As the local soledad component's cancel_bootstrap() method is called before
the close() method, we'd better not unset the proxied object there, and let
only the close method do it. This commit fixes this by just removing the line
that unsets the proxied object in the wrong place.