[bug] fail gracefully if no pixelated modules present
currently, we're distributing the wheels for the pixelated modules under
downloads.leap.se.

bootstrap script tried to download the pixelated modules, but it fails
on python versions < 2.7.9, apparently.

as a workaround, I make the import of the pixelated modules a non-fatal
error by setting a flag, and doing the launching of the pix UA
conditional on a successful import.

- Related: #8009